requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==2.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.2
